2015 Flight Delays and Cancellations
The U.S. Department of Transportation's (DOT) Bureau of Transportation Statistics tracks the on-time performance of domestic flights operated by large air carriers. Summary information on the number of on-time, delayed, canceled, and diverted flights is published in DOT's monthly Air Travel Consumer Report and in this data set of 2015 flight delays and cancellations.
The html file includes my analysis on the flights delayed. I have tried to create a few graphs and visualizations.